// src/app/pages/vehicles/components/vehicles-table/vehicles.interface.ts
export interface VehiclesInterface {
   idvehicle?: number;
   company_idcompany?: number;
   company_company_idcompany?: string;
   model?: string;
   type?: string;
   km?: number;
   trade?: string;
   year?: number;
   number?: number;
   is_deleted?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}